fix: refactored rl folder into mlps
