--- src/brittle_star_project/rl/__init__.py
@@ -1,25 +1,0 @@
-from .DummyAgent import Network, Critic, Actor, AgentParams, Storage
-from .base import (
-    RLAlgorithm,
-    RLModel,
-    Transition,
-    create_model,
-    register_rl_model,
-    registered_model_types,
-)
-from .random_policy_model import RandomPolicyModel
-
-__all__ = [
-    "RLAlgorithm",
-    "RLModel",
-    "RandomPolicyModel",
-    "Transition",
-    "create_model",
-    "register_rl_model",
-    "registered_model_types",
-    "Network",
-    "Critic",
-    "Actor",
-    "AgentParams",
-    "Storage",
-]
